Testing

+++ Task2/main-server/src/main_server/infrastructure/postgresql/admin_repo.rs
@@ -22,7 +22,7 @@ pub async fn get_admin_access_level(db: &PgPool, user_id: i32) -> Result<i32, sq
 }
 
 pub async fn is_admin(db: &PgPool, user_id: i32) -> Result<bool, sqlx::Error> {
-    let row: (i32,) = sqlx::query_as("select count(*) from admin where user_id = $1")
+    let row: (i64,) = sqlx::query_as("select count(*) from admin where user_id = $1")
         .bind(user_id)
         .fetch_one(db)
         .await?;
